requirements.txt
Django==1.7.1
MySQL-python==1.2.5
djangorestframework==2.4.4
django-storages
boto
